fix: delete 로직에 조건 추가

+++ src/main/java/com/example/tily/image/ImageService.java
@@ -35,7 +35,9 @@ public class ImageService {
                 .orElseThrow(() -> new CustomException(ExceptionCode.USER_NOT_FOUND));
 
         String storageFileName = s3Service.uploadFile(multipartFile, FileFolder.USER_IMAGE);
-        s3Service.deleteFile(user.getImage()); // s3에 업로드하고, 기존에 있던것은 지운다
+        if(user.getImage() != null) {
+            s3Service.deleteFile(user.getImage()); // s3에 업로드하고, 기존에 있던것은 지운다
+        }
 
         user.updateImage(storageFileName); // user의 image 필드는 파일명을 가진다
     }
@@ -56,8 +58,9 @@ public class ImageService {
                 .orElseThrow(() -> new CustomException(ExceptionCode.ROADMAP_NOT_FOUND));
 
         String storageFileName = s3Service.uploadFile(multipartFile, FileFolder.ROADMAP_IMAGE);
-        s3Service.deleteFile(roadmap.getImage());
-
+        if(roadmap.getImage() != null) {
+            s3Service.deleteFile(roadmap.getImage());
+        }
         roadmap.updateImage(storageFileName);
     }
